pyproject.toml
[project]
name = "deep-dives"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]